requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
openai==1.30.1
pydantic==2.7.1
pydantic-core==2.18.2
sqlalchemy==2.0.30
python-dotenv==1.0.1
httpx==0.27.0
tqdm==4.67.1
